Move parseXML function to common location
